update manifests for OSDK v0.18.1
changed the `placeholder` namespace to `windows-machine-config-operator`
so as to help with ci and to make sure that the operator in installed in
correct namespace

also added windows-machine-config-op.package.yaml to point to the
operator version that we want OLM to run on the cluster. This is needed
because we will be using `run packagemanifests` to run the operator via
OLM. `run bundles` is not yet supported by the operator-sdk.